Moved free related functions from ft_handle_error.c to ft_clean_game.c
